+++ BranchSDK/BNCDeviceInfo.h
@@ -1,6 +1,8 @@
 //
 //  BNCDeviceInfo.h
-//  Branch-TestBed
+//  BranchSDK
+//
+//  Class responsible for collating device information.
 //
 //  Created by Sojan P.R. on 3/22/16.
 //  Copyright © 2016 Branch Metrics. All rights reserved.
@@ -32,12 +34,12 @@
 @property (nonatomic, copy, readwrite) NSString *hardwareIdType;
 @property (nonatomic, assign, readwrite) BOOL isRealHardwareId;
 
+@property (nonatomic, copy, readwrite) NSString *anonId;
 @property (nonatomic, copy, readwrite) NSString *advertiserId;
 @property (nonatomic, copy, readwrite) NSString *vendorId;
 @property (nonatomic, copy, readwrite) NSString *optedInStatus;
 @property (nonatomic, assign, readwrite) BOOL isFirstOptIn;
 @property (nonatomic, assign, readwrite) BOOL isAdTrackingEnabled;
-@property (nonatomic, assign, readwrite) BOOL unidentifiedDevice;
 - (NSString *)localIPAddress;
 - (NSString *)connectionType;
 
@@ -61,7 +63,5 @@
 @property (nonatomic, copy, readwrite) NSString *pluginName;
 @property (nonatomic, copy, readwrite) NSString *pluginVersion;
 @property (nonatomic, copy, readwrite) NSString *branchSDKVersion;
-
-- (BOOL)isAppClip;
 
 @end