+++ BranchSDK/BNCDeviceInfo.m
@@ -1,10 +1,11 @@
 //
 //  BNCDeviceInfo.m
-//  Branch-TestBed
+//  BranchSDK
 //
 //  Created by Sojan P.R. on 3/22/16.
 //  Copyright © 2016 Branch Metrics. All rights reserved.
 //
+
 #import "BNCDeviceInfo.h"
 #import "BNCPreferenceHelper.h"
 #import "BNCSystemObserver.h"
@@ -31,9 +32,7 @@
 #pragma mark - BNCDeviceInfo
 
 @interface BNCDeviceInfo()
-
 @property (nonatomic, copy, readwrite) NSString *randomId;
-
 @end
 
 @implementation BNCDeviceInfo
@@ -62,6 +61,15 @@
     }
 }
 
+- (NSString *)loadAnonID {
+    NSString *tmp = [BNCPreferenceHelper sharedInstance].anonID;
+    if (!tmp) {
+        tmp = [NSUUID UUID].UUIDString;
+        [BNCPreferenceHelper sharedInstance].anonID = tmp;
+    }
+    return tmp;
+}
+
 - (void)loadDeviceInfo {
 
     BNCLocale *locale = [BNCLocale new];
@@ -70,60 +78,34 @@
     // The random id is regenerated per app launch.  This maintains existing behavior.
     self.randomId = [[NSUUID UUID] UUIDString];
     self.vendorId = [[UIDevice currentDevice].identifierForVendor UUIDString];
+    self.anonId = [self loadAnonID];
     [self checkAdvertisingIdentifier];
 
-    self.brandName = [BNCSystemObserver getBrand];
-    self.modelName = [BNCSystemObserver getModel];
-    self.osName = [BNCSystemObserver getOS];
-    self.osVersion = [BNCSystemObserver getOSVersion];
+    self.brandName = [BNCSystemObserver brand];
+    self.modelName = [BNCSystemObserver model];
+    self.osName = [BNCSystemObserver osName];
+    self.osVersion = [BNCSystemObserver osVersion];
     self.osBuildVersion = deviceSystem.systemBuildVersion;
 
     if (deviceSystem.cpuType) {
         self.cpuType = [deviceSystem.cpuType stringValue];
     }
 
-    self.screenWidth = [BNCSystemObserver getScreenWidth];
-    self.screenHeight = [BNCSystemObserver getScreenHeight];
-    self.screenScale = @([UIScreen mainScreen].scale);
+    self.screenWidth = [BNCSystemObserver screenWidth];
+    self.screenHeight = [BNCSystemObserver screenHeight];
+    self.screenScale = [BNCSystemObserver screenScale];
 
     #if !TARGET_OS_TV
     BNCTelephony *telephony = [BNCTelephony new];
     self.carrierName = telephony.carrierName;
-    #endif
+    #endif 
 
     self.locale = [NSLocale currentLocale].localeIdentifier;
     self.country = [locale country];
     self.language = [locale language];
-    self.environment = [self environment];
+    self.environment = [BNCSystemObserver environment];
     self.branchSDKVersion = [NSString stringWithFormat:@"ios%@", BNC_SDK_VERSION];
-    self.applicationVersion = [NSBundle mainBundle].infoDictionary[@"CFBundleShortVersionString"];
-    if (!self.applicationVersion.length) {
-        self.applicationVersion = [NSBundle mainBundle].infoDictionary[@"CFBundleVersionKey"];
-    }
-}
-
-// App Clips have a zero'd out IDFV
-- (BOOL)isAppClip {
-    if ([@"00000000-0000-0000-0000-000000000000" isEqualToString:[[UIDevice currentDevice].identifierForVendor UUIDString]]) {
-        return YES;
-    }
-    return NO;
-}
-
-- (NSString *)environment {
-    NSString *result = @"FULL_APP";
-    
-    if ([self isAppClip]) {
-        result = @"APP_CLIP";
-    }
-    
-    // iMessage has an extension id set in the Bundle
-    NSString *extensionType = [NSBundle mainBundle].infoDictionary[@"NSExtension"][@"NSExtensionPointIdentifier"];
-    if ([extensionType isEqualToString:@"com.apple.identitylookup.message-filter"]) {
-        result = @"IMESSAGE_APP";
-    }
-    
-    return result;
+    self.applicationVersion = [BNCSystemObserver applicationVersion];
 }
 
 - (NSString *)localIPAddress {
@@ -154,8 +136,8 @@
         self.isFirstOptIn = NO;
     }
     
-    self.isAdTrackingEnabled = [BNCSystemObserver adTrackingSafe];
-    self.advertiserId = [BNCSystemObserver getAdId];
+    self.isAdTrackingEnabled = [BNCSystemObserver adTrackingEnabled];
+    self.advertiserId = [BNCSystemObserver advertiserIdentifier];
     BOOL ignoreIdfa = [BNCPreferenceHelper sharedInstance].isDebug;
 
     if (self.advertiserId && !ignoreIdfa) {
@@ -191,6 +173,7 @@
             [dictionary bnc_safeSetObject:self.vendorId forKey:@"idfv"];
             [dictionary bnc_safeSetObject:self.advertiserId forKey:@"idfa"];
         }
+        [dictionary bnc_safeSetObject:[self anonId] forKey:@"anon_id"];
         [dictionary bnc_safeSetObject:[self localIPAddress] forKey:@"local_ip"];
 
         [dictionary bnc_safeSetObject:[self optedInStatus] forKey:@"opted_in_status"];