+++ BranchSDK/BNCPreferenceHelper.m
@@ -49,10 +49,16 @@ static NSString * const BRANCH_PREFS_KEY_ANALYTICS_MANIFEST = @"bnc_branch_analy
 static NSString * const BRANCH_PREFS_KEY_REFERRER_GBRAID = @"bnc_referrer_gbraid";
 static NSString * const BRANCH_PREFS_KEY_REFERRER_GBRAID_WINDOW = @"bnc_referrer_gbraid_window";
 static NSString * const BRANCH_PREFS_KEY_REFERRER_GBRAID_INIT_DATE = @"bnc_referrer_gbraid_init_date";
+static NSString * const BRANCH_PREFS_KEY_REFERRER_GCLID = @"bnc_referrer_gclid";
 static NSString * const BRANCH_PREFS_KEY_SKAN_CURRENT_WINDOW = @"bnc_skan_current_window";
 static NSString * const BRANCH_PREFS_KEY_FIRST_APP_LAUNCH_TIME = @"bnc_first_app_launch_time";
 static NSString * const BRANCH_PREFS_KEY_SKAN_HIGHEST_CONV_VALUE_SENT = @"bnc_skan_send_highest_conv_value";
 static NSString * const BRANCH_PREFS_KEY_SKAN_INVOKE_REGISTER_APP = @"bnc_invoke_register_app";
+                                                                
+static NSString * const BRANCH_PREFS_KEY_REFFERING_URL_QUERY_PARAMETERS = @"bnc_referring_url_query_parameters";
+
+static NSString * const BRANCH_PREFS_KEY_LOG_IAP_AS_EVENTS = @"bnc_log_iap_as_events";
+
 
 NSURL* /* _Nonnull */ BNCURLForBranchDirectory_Unthreaded(void);
 
@@ -72,37 +78,41 @@ NSURL* /* _Nonnull */ BNCURLForBranchDirectory_Unthreaded(void);
 
 @implementation BNCPreferenceHelper
 
+// since we override both setter and getter, these properties do not auto synthesize
 @synthesize
-            lastRunBranchKey = _lastRunBranchKey,
-            appVersion = _appVersion,
-            randomizedDeviceToken = _randomizedDeviceToken,
-            sessionID = _sessionID,
-            spotlightIdentifier = _spotlightIdentifier,
-            randomizedBundleToken = _randomizedBundleToken,
-            linkClickIdentifier = _linkClickIdentifier,
-            userUrl = _userUrl,
-            userIdentity = _userIdentity,
-            sessionParams = _sessionParams,
-            installParams = _installParams,
-            universalLinkUrl = _universalLinkUrl,
-            initialReferrer = _initialReferrer,
-            localUrl = _localUrl,
-            externalIntentURI = _externalIntentURI,
-            isDebug = _isDebug,
-            retryCount = _retryCount,
-            retryInterval = _retryInterval,
-            timeout = _timeout,
-            lastStrongMatchDate = _lastStrongMatchDate,
-            checkedFacebookAppLinks = _checkedFacebookAppLinks,
-            checkedAppleSearchAdAttribution = _checkedAppleSearchAdAttribution,
-            appleSearchAdDetails = _appleSearchAdDetails,
-            requestMetadataDictionary = _requestMetadataDictionary,
-            instrumentationDictionary = _instrumentationDictionary,
-            referrerGBRAID = _referrerGBRAID,
-            referrerGBRAIDValidityWindow = _referrerGBRAIDValidityWindow,
-            skanCurrentWindow = _skanCurrentWindow,
-            firstAppLaunchTime = _firstAppLaunchTime,
-            highestConversionValueSent = _highestConversionValueSent;
+    lastRunBranchKey = _lastRunBranchKey,
+    appVersion = _appVersion,
+    randomizedDeviceToken = _randomizedDeviceToken,
+    sessionID = _sessionID,
+    spotlightIdentifier = _spotlightIdentifier,
+    randomizedBundleToken = _randomizedBundleToken,
+    linkClickIdentifier = _linkClickIdentifier,
+    userUrl = _userUrl,
+    userIdentity = _userIdentity,
+    sessionParams = _sessionParams,
+    installParams = _installParams,
+    universalLinkUrl = _universalLinkUrl,
+    initialReferrer = _initialReferrer,
+    localUrl = _localUrl,
+    externalIntentURI = _externalIntentURI,
+    isDebug = _isDebug,
+    retryCount = _retryCount,
+    retryInterval = _retryInterval,
+    timeout = _timeout,
+    lastStrongMatchDate = _lastStrongMatchDate,
+    checkedFacebookAppLinks = _checkedFacebookAppLinks,
+    checkedAppleSearchAdAttribution = _checkedAppleSearchAdAttribution,
+    appleSearchAdDetails = _appleSearchAdDetails,
+    requestMetadataDictionary = _requestMetadataDictionary,
+    instrumentationDictionary = _instrumentationDictionary,
+    referrerGBRAID = _referrerGBRAID,
+    referrerGBRAIDValidityWindow = _referrerGBRAIDValidityWindow,
+    skanCurrentWindow = _skanCurrentWindow,
+    firstAppLaunchTime = _firstAppLaunchTime,
+    highestConversionValueSent = _highestConversionValueSent,
+    logInAppPurchasesAsBranchEvents = _logInAppPurchasesAsBranchEvents,
+    referringURLQueryParameters = _referringURLQueryParameters,
+    anonID = _anonID;
 
 + (BNCPreferenceHelper *)sharedInstance {
     static BNCPreferenceHelper *preferenceHelper;
@@ -238,6 +248,20 @@ NSURL* /* _Nonnull */ BNCURLForBranchDirectory_Unthreaded(void);
     if (randomizedDeviceToken == nil || ![_randomizedDeviceToken isEqualToString:randomizedDeviceToken]) {
         _randomizedDeviceToken = randomizedDeviceToken;
         [self writeObjectToDefaults:BRANCH_PREFS_KEY_RANDOMIZED_DEVICE_TOKEN value:randomizedDeviceToken];
+    }
+}
+
+- (NSString *)anonID {
+    if (!_anonID) {
+        _anonID = [self readStringFromDefaults:@"bnc_anon_id"];
+    }
+    return _anonID;
+}
+
+- (void)setAnonID:(NSString *)anonID {
+    if (![_anonID isEqualToString:anonID]) {
+        _anonID = anonID;
+        [self writeObjectToDefaults:@"bnc_anon_id" value:anonID];
     }
 }
 
@@ -674,6 +698,23 @@ NSURL* /* _Nonnull */ BNCURLForBranchDirectory_Unthreaded(void);
     }
 }
 
+- (void)setReferringURLQueryParameters:(NSMutableDictionary *)parameters {
+    @synchronized(self) {
+        _referringURLQueryParameters = parameters;
+        [self writeObjectToDefaults:BRANCH_PREFS_KEY_REFFERING_URL_QUERY_PARAMETERS value:parameters];
+    }
+}
+
+- (NSMutableDictionary *)referringURLQueryParameters {
+    @synchronized(self) {
+        if (!_referringURLQueryParameters) {
+            _referringURLQueryParameters = (NSMutableDictionary *)[self readObjectFromDefaults:BRANCH_PREFS_KEY_REFFERING_URL_QUERY_PARAMETERS];
+        }
+    }
+    return _referringURLQueryParameters;
+}
+
+
 - (NSString *) referrerGBRAID {
     @synchronized(self) {
         if (!_referrerGBRAID) {
@@ -783,6 +824,15 @@ NSURL* /* _Nonnull */ BNCURLForBranchDirectory_Unthreaded(void);
     }
 }
 
+- (BOOL) logInAppPurchasesAsBranchEvents {
+    _logInAppPurchasesAsBranchEvents = [self readBoolFromDefaults:BRANCH_PREFS_KEY_LOG_IAP_AS_EVENTS];
+    return _logInAppPurchasesAsBranchEvents;
+}
+
+- (void) setLogInAppPurchasesAsBranchEvents:(BOOL)logInAppPurchasesAsBranchEvents {
+    _logInAppPurchasesAsBranchEvents = logInAppPurchasesAsBranchEvents;
+    [self writeBoolToDefaults:BRANCH_PREFS_KEY_LOG_IAP_AS_EVENTS value:logInAppPurchasesAsBranchEvents];
+}
 
 - (void) clearTrackingInformation {
     @synchronized(self) {
@@ -807,6 +857,8 @@ NSURL* /* _Nonnull */ BNCURLForBranchDirectory_Unthreaded(void);
         self.requestMetadataDictionary = nil;
         self.lastStrongMatchDate = nil;
         self.userIdentity = nil;
+        self.referringURLQueryParameters = nil;
+        self.anonID = nil;
     }
 }