+++ BranchSDK/Branch.m
@@ -46,6 +46,8 @@
 #import "NSError+Branch.h"
 #import "BNCLog.h"
 #import "UIViewController+Branch.h"
+#import <StoreKit/StoreKit.h>
+#import "BNCReferringURLUtility.h"
 
 #if !TARGET_OS_TV
 #import "BNCUserAgentCollector.h"
@@ -124,7 +126,7 @@ typedef NS_ENUM(NSInteger, BNCInitStatus) {
     BNCInitStatusInitialized
 };
 
-@interface Branch() <BranchDeepLinkingControllerCompletionDelegate> {
+@interface Branch() <BranchDeepLinkingControllerCompletionDelegate, SKPaymentTransactionObserver> {
     NSInteger _networkCount;
     BNCURLFilter *_userURLFilter;
 }
@@ -151,6 +153,11 @@ typedef NS_ENUM(NSInteger, BNCInitStatus) {
 #endif
 
 @property (nonatomic, copy, nullable) void (^sceneSessionInitWithCallback)(BNCInitSessionResponse * _Nullable initResponse, NSError * _Nullable error);
+
+// Support for deferred SDK initialization. Used to support slow plugin runtime startup.
+// This is enabled by setting deferInitForPluginRuntime to true in branch.json
+@property (nonatomic, assign, readwrite) BOOL deferInitForPluginRuntime;
+@property (nonatomic, copy, nullable) void (^cachedInitBlock)(void);
 
 @end
 
@@ -226,13 +233,14 @@ typedef NS_ENUM(NSInteger, BNCInitStatus) {
     [self loadUserAgent];
     
     BranchJsonConfig *config = BranchJsonConfig.instance;
+    self.deferInitForPluginRuntime = config.deferInitForPluginRuntime;
+    
+    if (config.enableLogging) {
+        [self enableLogging];
+    }
     
     if (config.checkPasteboardOnInstall) {
         [self checkPasteboardOnInstall];
-    }
-    
-    if (config.delayInitToCheckForSearchAds) {
-        [self delayInitToCheckForSearchAds];
     }
 
     if (config.enableFacebookLinkCheck) {
@@ -535,7 +543,7 @@ static NSString *bnc_branchKey = nil;
 
 + (void)setReferrerGbraidValidityWindow:(NSTimeInterval)validityWindow{
     @synchronized(self) {
-        [BNCPreferenceHelper sharedInstance].referrerGBRAIDValidityWindow = validityWindow;
+        [BNCPreferenceHelper sharedInstance].referringURLQueryParameters[BRANCH_REQUEST_KEY_REFERRER_GBRAID][BRANCH_URL_QUERY_PARAMETERS_VALIDITY_WINDOW_KEY] = @(validityWindow);
     }
 }
 
@@ -613,8 +621,10 @@ static NSString *bnc_branchKey = nil;
 
 - (void)initSceneSessionWithLaunchOptions:(NSDictionary *)options isReferrable:(BOOL)isReferrable explicitlyRequestedReferrable:(BOOL)explicitlyRequestedReferrable automaticallyDisplayController:(BOOL)automaticallyDisplayController
                   registerDeepLinkHandler:(void (^)(BNCInitSessionResponse * _Nullable initResponse, NSError * _Nullable error))callback {
-    self.sceneSessionInitWithCallback = callback;
-    [self initSessionWithLaunchOptions:options isReferrable:isReferrable explicitlyRequestedReferrable:explicitlyRequestedReferrable automaticallyDisplayController:automaticallyDisplayController];
+    [self deferInitBlock:^{
+        self.sceneSessionInitWithCallback = callback;
+        [self initSessionWithLaunchOptions:options isReferrable:isReferrable explicitlyRequestedReferrable:explicitlyRequestedReferrable automaticallyDisplayController:automaticallyDisplayController];
+    }];
 }
 
 - (void)initSessionWithLaunchOptions:(NSDictionary *)options
@@ -703,17 +713,9 @@ static NSString *bnc_branchKey = nil;
     // this allows foreground links to callback
     self.initializationStatus = BNCInitStatusUninitialized;
 
-    //check the referring url/uri for query parameter gbraid
-    NSURLComponents *components = [NSURLComponents componentsWithURL:url resolvingAgainstBaseURL:NO];
-    NSString *gbraidValue = nil;
-    for(NSURLQueryItem *item in components.queryItems){
-        if([item.name isEqualToString:@"gbraid"])
-            gbraidValue = item.value;
-    }
-    
-    if (gbraidValue) {
-        self.preferenceHelper.referrerGBRAID = gbraidValue;
-    }
+    //Check the referring url/uri for query parameters and save them
+    BNCReferringURLUtility *utility = [BNCReferringURLUtility new];
+    [utility parseReferringURL:url];
     
     NSString *pattern = nil;
     pattern = [self.urlFilter patternMatchingURL:url];
@@ -1675,6 +1677,10 @@ static NSString *bnc_branchKey = nil;
                     cache:[[BNCLinkCache alloc] init]
                     preferenceHelper:preferenceHelper
                     key:key];
+            
+            if ([BNCPreferenceHelper sharedInstance].logInAppPurchasesAsBranchEvents == YES) {
+                [[SKPaymentQueue defaultQueue] addTransactionObserver:branch];
+            }
         });
         return branch;
     }
@@ -2136,9 +2142,36 @@ static inline void BNCPerformBlockOnMainThreadSync(dispatch_block_t block) {
 
 #pragma mark - Session Initialization
 
+// Defers block until notifyNativeToInit is called.
+- (BOOL)deferInitBlock:(void (^)(void))block {
+    BOOL deferred = NO;
+    @synchronized (self) {
+        if (self.deferInitForPluginRuntime) {
+            self.cachedInitBlock = block;
+            deferred = YES;
+        }
+    }
+    
+    if (!deferred && block) {
+        block();
+    }
+    return deferred;
+}
+
+// Releases deferred init block
+- (void)notifyNativeToInit {
+    @synchronized (self) {
+        self.deferInitForPluginRuntime = NO;
+    }
+    
+    if (self.cachedInitBlock) {
+        self.cachedInitBlock();
+    }
+    self.cachedInitBlock = nil;
+}
+
 // SDK-631 Workaround to maintain existing error handling behavior.
 // Some methods require init before they are called.  Instead of returning an error, we try to fix the situation by calling init ourselves.
-// There is a follow up ticket to improve this.  SDK-633
 - (void)initSafetyCheck {
     if (self.initializationStatus == BNCInitStatusUninitialized) {
         BNCLogDebug(@"Branch avoided an error by preemptively initializing.");
@@ -2147,6 +2180,16 @@ static inline void BNCPerformBlockOnMainThreadSync(dispatch_block_t block) {
 }
 
 - (void)initUserSessionAndCallCallback:(BOOL)callCallback sceneIdentifier:(NSString *)sceneIdentifier {
+    
+    // ignore lifecycle calls while waiting for a plugin runtime.
+    @synchronized (self) {
+        if (self.deferInitForPluginRuntime) {
+            //NSString *debug = [NSString stringWithFormat:@"Init is deferred, ignoring call: %@", NSThread.callStackSymbols];
+            //BNCLogDebug(debug);
+            return;
+        }
+    }
+    
     dispatch_async(self.isolationQueue, ^(){
         NSString *urlstring = nil;
         if (self.preferenceHelper.universalLinkUrl.length) {
@@ -2557,6 +2600,53 @@ static inline void BNCPerformBlockOnMainThreadSync(dispatch_block_t block) {
     [[BNCServerRequestQueue getInstance] clearQueue];
     [BranchOpenRequest releaseOpenResponseLock];
     [BNCPreferenceHelper clearAll];
+}
+
+#pragma mark - SKPaymentTransactionObserver Methods
++ (void)setLogInAppPurchasesAsEventsEnabled:(BOOL)enabled {
+    @synchronized(self) {
+        if (enabled) {
+            [BNCPreferenceHelper sharedInstance].logInAppPurchasesAsBranchEvents = YES;
+        } else {
+            [BNCPreferenceHelper sharedInstance].logInAppPurchasesAsBranchEvents = NO;
+        }
+    }
+}
+
++ (BOOL)logInAppPurchasesBranchEventsEnabled {
+    @synchronized(self) {
+        return [BNCPreferenceHelper sharedInstance].logInAppPurchasesAsBranchEvents;
+    }
+}
+
+//Logs incoming in-app purchases as events
+- (void)paymentQueue:(SKPaymentQueue *)queue updatedTransactions:(NSArray<SKPaymentTransaction *> *)transactions {
+    for (SKPaymentTransaction *transaction in transactions) {
+        switch (transaction.transactionState) {
+            case SKPaymentTransactionStatePurchased: {
+                
+                [[SKPaymentQueue defaultQueue] finishTransaction:(SKPaymentTransaction *)transaction];
+                
+                if ([BNCPreferenceHelper sharedInstance].logInAppPurchasesAsBranchEvents == YES) {
+                    BNCLogDebug([NSString stringWithFormat:@"Automatically logging transaction as Branch event."]);
+                    
+                    BranchEvent *event = [BranchEvent standardEvent:BranchStandardEventPurchase];
+                    [event logEventWithTransaction:transaction];
+                }
+                break;
+            }
+            case SKPaymentTransactionStateFailed: {
+                [[SKPaymentQueue defaultQueue] finishTransaction:(SKPaymentTransaction *)transaction];
+                break;
+            }
+            case SKPaymentTransactionStateRestored: {
+                [[SKPaymentQueue defaultQueue] finishTransaction:(SKPaymentTransaction *)transaction];
+                break;
+            }
+            default:
+                break;
+        }
+    }
 }
 
 @end